Move osdf-client config_mgr into the pelican CLI
