+++ namespaces/resources/namespaces.json
